\name{Carnivores}
\docType{data}
\alias{Carnivores}
\title{Presence of terrestrial carnivores of the family Felidae and environmental variables.}
\description{Presence of terrestrial carnivores of the family Felidae and mean of environmental variables in cells of 1 degree x 1 degree around the world.}
\usage{data(Carnivores)}
\format{A matrix of the presence of terrestrial carnivores of the family Felidae and the mean altitude, mean annual temperature, isothermality, temperature seasonality and mean annual precipitation in cells of 1 degree x 1 degree around the world.}
\source{The range maps of the species were obtained from the International Union for Conservation of Nature (IUCN) at the web page http://www.iucn.org/. The data of the mean annual temperature (BIO1), isothermality (BIO3), temperature seasonality (BIO4) and mean annual precipitation (BIO12) were downloaded from the web http://www.worldclim.org/. Both range maps and environmental variables were inputted into ModestR (www.ipez.es/ModestR) and the output file from ModestR is a CSV file that was converted to a RData file.}
\references{
García-Roselló, E., Guisande, C., González-Dacosta, J., Heine, J., Pelayo-Villamil, P., Manjarrés-Hernández, A., Vaamonde, A. & Granado-Lorencio, C. (2013) ModestR: a software tool for managing and analyzing species distribution map databases. \emph{Ecography}, 36, 1202-1207.

Hijmans, R.J., Cameron, S.E., Parra, J.L., Jones, P.G. and Jarvis, A. (2005) Very high resolution interpolated climate surfaces for global land areas. \emph{International Journal of Climatology}, 25, 1965-1978.

IUCN (2012) The IUCN Red List of Threatened Species. Version 2012.2. <http://www.iucnredlist.org>. Downloaded on 17 October 2012.

Pelayo-Villamil, P., Guisande, C., González-Vilas. L., Carvajal-Quintero. J.D., Jiménez-Segura, L.F., García-Roselló, E., Heine, J., González-Dacosta, J., Manjarrés-Hernández, A., Vaamonde, A., Granado-Lorencio, C. (2012) ModestR: Una herramienta infromática para el estudio de los ecosistemas acuáticos de Colombia. \emph{Actualidades Biológicas}, 34, 225-239.
}
\keyword{Carnivores}

 
