\name{Polar.coordinates}
\docType{data}
\encoding{latin1}
\alias{Polar.coordinates}
\title{Polar cooordinates obtained from the function \link{Niche}}
\description{Polar cooordinates.CSV file obtained from the function \link{Niche} using presence data of terrestrial carnivores of the family Felidae and mean of environmental variables in cells of 1 degree x 1 degree around the world.}
\usage{data(Polar.coordinates)}
\format{A matrix with the taxonomy, polar coordinates and mean values of the environemntal variables for each polar coordinate.}
\source{The range maps of the species were obtained from the International Union for Conservation of Nature (IUCN) at the web page \url{https://www.iucn.org/}. The data of the mean annual temperature (BIO1), isothermality (BIO3), temperature seasonality (BIO4) and mean annual precipitation (BIO12) were downloaded from the web \url{https://www.worldclim.org/}. Both range maps and environmental variables were inputted into ModestR and the output file was used in the function Niche() and the file Polar coordinates.CSV obtained was converted to a RData file.}
\references{
García-Roselló, E., Guisande, C., González-Dacosta, J., Heine, J., Pelayo-Villamil, P., Manjarrés-Hernández, A., Vaamonde, A. & Granado-Lorencio, C. (2013) ModestR: a software tool for managing and analyzing species distribution map databases. \emph{Ecography}, 36, 1202-1207.

Hijmans, R.J., Cameron, S.E., Parra, J.L., Jones, P.G. and Jarvis, A. (2005) Very high resolution interpolated climate surfaces for global land areas. \emph{International Journal of Climatology}, 25, 1965-1978.

IUCN (2012) The IUCN Red List of Threatened Species. Version 2012.2. \url{https://www.iucnredlist.org}. Downloaded on 17 October 2012.

Pelayo-Villamil, P., Guisande, C., González-Vilas. L., Carvajal-Quintero. J.D., Jiménez-Segura, L.F., García-Roselló, E., Heine, J., González-Dacosta, J., Manjarrés-Hernández, A., Vaamonde, A., Granado-Lorencio, C. (2012) ModestR: Una herramienta infromática para el estudio de los ecosistemas acuáticos de Colombia. \emph{Actualidades Biológicas}, 34, 225-239.
}
\keyword{Polar.coordinates}

 
