\name{Carnivores}
\docType{data}
\alias{Carnivores}
\title{Presence of terrestrial carnivores of the family Felidae and environmental variables.}
\description{Presence of terrestrial carnivores of the family Felidae and mean of environmental
 variables in cells of 1 degree x 1 degree around the world.}
\usage{data(Carnivores)}
\format{A matrix of the presence of terrestrial carnivores of the family
 Felidae and the mean altitude, mean temperature and mean precipitation
 in cells of 1 degree x 1 degree around the world.}
\source{The range maps of the species were obtained from the International
 Union for Conservation of Nature (IUCN) at the web page http://www.iucn.org/.
 The data of the mean temperature (BIO1) and the mean precipitation (BIO12),
 from 1950 to 2000, were downloaded from the web http://www.worldclim.org/.
 Both range maps and environmental variables were input into ModestR
 (www.ipez.es/ModestR) and the output file from ModestR is a CSV file that was
 converted to a Rdat.}
\keyword{Carnivores}

 
